deleted the header component, the header is now in the existing app-header component
